correcoes finais

+++ _controller/ajax/delete_animal.php
@@ -4,14 +4,23 @@ include_once '../config.php';
 
 $id = $_POST['id'];
 
-$delete_animal = "delete from ani_animal
+$delete_atendimento = "delete from atd_atendimento
+                       WHERE atd_id_animal = $id";
+
+if(mysqli_query(connect(),$delete_atendimento)){
+    $delete_animal = "delete from ani_animal
                   where ani_id_animal = $id";
 
-if(mysqli_query(connect(),$delete_animal)){
+    if(mysqli_query(connect(),$delete_animal)){
 
-    echo 'Agendamento deletado com sucesso!';
-    exit;
+        echo '1';
+        exit;
 
+    }else{
+        echo '2';
+    }
 }else{
-    echo 'Erro ao deletar agendamento, entre em contato com o administrador do sistema!';
+    echo '3';
 }
+
+
